Fixed various bugs
The Ninja Airports API swapped their plans from limited free to premium exclusive, so I had to swap out every code and part that used it. With the implementation of the new AeroDataBox API i was able to add a feature that lets the user search for airports based on city names. Also fixed the problem with the logo.